+++ doc/includes/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
+## 0.9.4
+
+### What's new
+
+  * Fixed [#627](https://github.com/Vincit/objection.js/issues/627)
+  * Fixed [#671](https://github.com/Vincit/objection.js/issues/671)
+  * Fixed [#672](https://github.com/Vincit/objection.js/issues/672)
+  * Fixed [#674](https://github.com/Vincit/objection.js/issues/674)
+
 ## 0.9.3
 
+### What's new
+
   * Add beforeInsert hook for relations. [#649](https://github.com/Vincit/objection.js/issues/649) [#19](https://github.com/Vincit/objection.js/issues/19)
+  * Add [`relatedFindQueryMutates`](#relatedfindquerymutates) and [`relatedInsertQueryMutates`](#relatedinsertquerymutates) configs as well as [`$setRelated`](#_s_setrelated) and [`$appendRelated`](#_s_appendrelated) helpers. [#599](https://github.com/Vincit/objection.js/issues/599)
   * Fixed [#648](https://github.com/Vincit/objection.js/issues/648)
 
 ## 0.9.2